requirements.txt
streamlit
transformers
torch
PyPDF2
langchain
faiss-cpu
google-generativeai
pymupdf
spacy
pyrebase4
streamlit-option-menu
